+++ docs/release-notes.rst
@@ -7,4 +7,8 @@ The server is deployed within a Docker image in the Docker registry.
 Version 1.0.0 December 2017
 ---------------------------
 
-* Supports the above three end points
+* Supports the three end points: /putBlueprint, /{operation}, /putDockerInfo
+
+Version 1.0.1 Feburary 2018
+
+* Supports new blueprint JSON format which contains node_type, message_name, and proto_url